pyproject.toml
[project]
name = "kslack"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "slack-bolt>=1.26.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "prometheus-client>=0.20.0",
    "structlog>=24.0.0",
    "aiohttp>=3.9.0",
    "a2a-sdk>=0.2.16,<0.4.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "types-pyyaml",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
]

[tool.ruff]
target-version = "py313"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "W", "B", "Q", "I", "ASYNC", "T20"]
ignore = ["F401", "E501", "B008", "ASYNC109"]

[tool.mypy]
python_version = "3.13"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = "test_*.py"
python_functions = "test_*"
log_cli = true
log_cli_level = "INFO"
testpaths = ["tests"]
